backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "dima_in_the_forest", email = "da.kutsev@gmail.com"},
    {name = "dima_in_the_forest", email = "da.kutsev@gmail.cpm"},
]
dependencies = [
    "openai>=1.10.0",
    "python-decouple>=3.8",
    "python-multipart>=0.0.6",
    "requests==2.28.2",
    "fastapi>=0.109.0",
    "uvicorn[standart]>=0.27.0",
    "loguru>=0.7.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
